ShellPkg: Add ACPI data store
Add a store for useful ACPI data that can be used in validation. This
data will be collected from the parsers, stored in the data store and
then can be accessed by validators once all parsing is complete.

The data is stored dynamically as nodes in a linked list that is
accessed by META_DATA_TYPE.

Signed-off-by: Chris Jones <christopher.jones@arm.com>
Change-Id: Ie8bab3987723f83540bec50d08eee30f438aee90